Cocoa Port: - Change gpuColorFormat property data type from UInt32 to NSUInteger.

+++ desmume/src/cocoa/cocoa_GPU.h
@@ -37,7 +37,7 @@ class GPUEventHandlerOSX;
 @property (assign) UInt32 gpuStateFlags;
 @property (assign) NSSize gpuDimensions;
 @property (assign) NSUInteger gpuScale;
-@property (assign) UInt32 gpuColorFormat;
+@property (assign) NSUInteger gpuColorFormat;
 @property (readonly) pthread_rwlock_t *gpuFrameRWLock;
 
 @property (assign) BOOL layerMainGPU;
